#1 [add] basic function
